Revert "Plan 9 from Bell Labs Third Edition 2000-06-07"
This reverts commit 7dd7cddf99dd7472612f1413b4da293630e6b1bc.

+++ sys/src/cmd/astro/cosadd.c
@@ -2,7 +2,7 @@
 
 
 void
-icosadd(double *fp, char *cp)
+icosadd(float *fp, char *cp)
 {
 
 	cafp = fp;
